moved to root & works in both directories

+++ createComponent.js
@@ -2,14 +2,14 @@
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
-const inquirer = require('inquirer');
-const fs = require('fs');
+const inquirer = require("inquirer");
+const fs = require("fs");
 
 const copyrightHeader =
-  '/*---------------------------------------------------------------------------------------------\n ' +
-  '* Copyright (c) Bentley Systems, Incorporated. All rights reserved.\n ' +
-  '* See LICENSE.md in the project root for license terms and full copyright notice.\n ' +
-  '*--------------------------------------------------------------------------------------------*/';
+  "/*---------------------------------------------------------------------------------------------\n " +
+  "* Copyright (c) Bentley Systems, Incorporated. All rights reserved.\n " +
+  "* See LICENSE.md in the project root for license terms and full copyright notice.\n " +
+  "*--------------------------------------------------------------------------------------------*/";
 
 const makeDir = (path) => {
   if (!fs.existsSync(path)) {
@@ -18,21 +18,23 @@ const makeDir = (path) => {
 };
 
 const writeFile = (obj) => {
-  fs.writeFile(obj.path, obj.template, (error) => {
-    if (error) {
-      throw new Error(error.message);
-    }
-    console.log('Successfully wrote', obj.path);
-  });
+  if (!fs.existsSync(obj.path)) {
+    fs.writeFile(obj.path, obj.template, (error) => {
+      if (error) {
+        throw new Error(error.message);
+      }
+      console.log("Successfully wrote", obj.path);
+    });
+  }
 };
 
 const appendFile = (obj) => {
   if (fs.existsSync(obj.path)) {
-    let file = fs.readFileSync(obj.path, 'utf-8');
+    let file = fs.readFileSync(obj.path, "utf-8");
     let tester = new RegExp(`.*${obj.template}.*`);
     let count = 0;
 
-    for (const line of file.split('\n')) {
+    for (const line of file.split("\n")) {
       if (tester.test(line)) {
         throw new Error(`Duplicate line in ${obj.path}: ${count}`);
       }
@@ -43,7 +45,7 @@ const appendFile = (obj) => {
       if (error) {
         throw new Error(error.message);
       }
-      console.log('Successfully appended', obj.path);
+      console.log("Successfully appended", obj.path);
     });
   } else {
     writeFile(obj);
@@ -51,7 +53,7 @@ const appendFile = (obj) => {
 };
 
 const functionalTemplate = (
-  name,
+  name
 ) => `export const ${name} = (props: ${name}Props) => {
   const { ...rest } = props;
   useTheme();
@@ -85,7 +87,7 @@ const storiesFactory = (directory, name) => {
     template: `${copyrightHeader}
 import { Story, Meta } from '@storybook/react';
 import React from 'react';
-import { ${name}, ${name}Props } from '../../../packages/iTwinUI-react/src/core';
+import { ${name}, ${name}Props } from '@itwin/itwinui-react';
 
 export default {
   component: ${name},
@@ -160,7 +162,7 @@ it('should be improved', () => {
 
 const indexAppendFactory = (directory, name) => {
   return {
-    path: `${directory.split('/').slice(0, -1).join('/')}/index.ts`,
+    path: `${directory.split("/").slice(0, -1).join("/")}/index.ts`,
     template: `
 export { ${name} } from './${name}';
 export type { ${name}Props } from './${name}';
@@ -171,17 +173,50 @@ export type { ${name}Props } from './${name}';
 inquirer
   .prompt([
     {
-      name: 'component',
-      type: 'input',
-      message: 'What is the name of the component?',
+      name: "component",
+      type: "input",
+      message: "What is the name of the component?",
     },
   ])
   .then(({ component }) => {
-    console.log();
-    let directory = 'packages/iTwinUI-react/src/core';
+    //   let directory = "packages/iTwinUI-react/src/core";
+    //   let levels = [];
+
+    //   directory += `/${component}`;
+    //   levels.push(directory);
+
+    //   let componentOut = componentFactory(directory, component);
+    //   let componentTest = componentTestFactory(directory, component);
+    //   let componentIndex = componentIndexFactory(directory, component);
+    //   let indexAppend = indexAppendFactory(directory, component);
+
+    //   for (const level of levels) {
+    //     makeDir(level);
+    //   }
+
+    //   appendFile(indexAppend);
+
+    //   writeFile(componentOut);
+    //   writeFile(componentTest);
+    //   writeFile(componentIndex);
+
+    //   writeFile(storiesFactory("apps/storybook/src", component));
+    //   writeFile(storyTestsFactory("apps/storybook/src", component));
+
+    const currentDirectory = process.cwd();
+    const isInPackage = currentDirectory.includes("packages");
+
+    let directory = "";
+    let storyDirectory = "";
     let levels = [];
 
-    directory += `/${component}`;
+    if (isInPackage) {
+      directory = `src/core/${component}`;
+      storyDirectory = "../../apps/storybook/src";
+    } else {
+      directory = `packages/iTwinUI-react/src/core/${component}`;
+      storyDirectory = "apps/storybook/src";
+    }
     levels.push(directory);
 
     let componentOut = componentFactory(directory, component);
@@ -199,6 +234,6 @@ inquirer
     writeFile(componentTest);
     writeFile(componentIndex);
 
-    writeFile(storiesFactory('apps/storybook/src', component));
-    writeFile(storyTestsFactory('apps/storybook/src', component));
+    writeFile(storiesFactory(storyDirectory, component));
+    writeFile(storyTestsFactory(storyDirectory, component));
   });
